[level 2] Title: 피보나치 수, Time: 1.62 ms, Memory: 72 MB -BaekjoonHub

+++ 프로그래머스/lv2/12945. 피보나치 수/피보나치 수.java
@@ -11,7 +11,7 @@ class Solution {
             answerBefore = temp;
         }
         
-        return (int)answer;
+        return answer;
         
     }
 }
